fix the flaws in lab1-1

+++ lab1-1/src/ring_buffer.c
@@ -1,6 +1,7 @@
+#include <stddef.h>
 #include "ring_buffer.h"
 
-void write(struct RingBuffer *buff, const unsigned char *str, unsigned int len_str)
+void write(struct RingBuffer *buff, const unsigned char *str, size_t len_str)
 {
     for (unsigned int i = 0; i < len_str; ++i)
     {
@@ -9,7 +10,7 @@ void write(struct RingBuffer *buff, const unsigned char *str, unsigned int len_s
     }
 }
 
-unsigned char get_elem(const struct RingBuffer *buff, unsigned int i)
+unsigned char get_elem(const struct RingBuffer *buff, size_t i)
 {
     return buff->data[i & buff->mask];
 }
